ref(general): Move filters into own module
