+++ Cargo.toml
@@ -34,6 +34,7 @@ lazy_static = "1.3.0"
 log = { version = "0.4.8", features = ["serde"] }
 pretty_env_logger = "0.3.0"
 relay-common = { path = "common" }
+relay-filter = { path = "filter" }
 relay-server = { path = "server" }
 relay-general = { path = "general" }
 sentry ={ version = "0.17.0", features = ["with_debug_meta"] }
