+++ filter/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "relay-filter"
+authors = ["Sentry <oss@sentry.io>"]
+description = "Sentry Event filters"
+homepage = "https://getsentry.github.io/relay/"
+repository = "https://github.com/getsentry/relay"
+version = "0.4.65"
+edition = "2018"
+license-file = "../LICENSE"
+
+[dependencies]
+globset = "0.4.4"
+ipnetwork = "0.14.0"
+lazy_static = "1.3.0"
+lazycell = "1.2.1"
+regex = "1.2.0"
+relay-general = { path = "../general" }
+serde = { version = "1.0.98", features = ["derive"] }
+
+[dev-dependencies]
+insta = "0.12.0"
+serde_json = "1.0.40"
