+++ filter/src/common.rs
@@ -142,22 +142,45 @@ impl<'de> Deserialize<'de> for GlobPatterns {
 /// (e.g.  "browser-extensions")
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Serialize, Hash)]
 pub enum FilterStatKey {
+    /// Filtered by ip address.
     IpAddress,
+
+    /// Filtered by release name (version).
     ReleaseVersion,
+
+    /// Filtered by error message.
     ErrorMessage,
+
+    /// Filtered by browser extension.
     BrowserExtensions,
+
+    /// Filtered by legacy browser version.
     LegacyBrowsers,
+
+    /// Filtered due to localhost restriction.
     Localhost,
+
+    /// Filtered as known web crawler.
     WebCrawlers,
+
+    /// Filtered due to invalid CSP policy.
     InvalidCsp,
-    // Cors, // NOTE: Although cors is in the Sentry's FilterStatKey enum it is used for
-    // Invalid outcomes and therefore should logically belong to OutcomeInvalidReason
-    // that is why it was commented here and moved to OutcomeInvalidReason enum
-    /// note: Not returned by any filters implemented in Rust.
-    DiscardedHash,
 }
 
+// An event grouped to a removed group.
+//
+// Not returned by any filters implemented in Rust.
+// DiscardedHash,
+
+// Invalid CORS header.
+//
+// NOTE: Although cors is in the Sentry's FilterStatKey enum it is used for
+// Invalid outcomes and therefore should logically belong to OutcomeInvalidReason
+// that is why it was commented here and moved to OutcomeInvalidReason enum
+// Cors,
+
 impl FilterStatKey {
+    /// Returns the string identifier of the filter stat key.
     pub fn name(self) -> &'static str {
         match self {
             FilterStatKey::IpAddress => "ip-address",
@@ -168,7 +191,6 @@ impl FilterStatKey {
             FilterStatKey::Localhost => "localhost",
             FilterStatKey::WebCrawlers => "web-crawlers",
             FilterStatKey::InvalidCsp => "invalid-csp",
-            FilterStatKey::DiscardedHash => "discarded-hash",
         }
     }
 }
