+++ filter/src/lib.rs
@@ -2,15 +2,15 @@
 //!
 //! Events may be filtered base on the following configurable criteria.
 //!
-//! * localhost ( filter events originating from the local machine)
-//! * browser extensions ( filter events caused by known problematic browser extensions)
-//! * web crawlers ( filter events sent by user agents known to be web crawlers)
-//! * legacy browsers ( filter events originating from legacy browsers, can be configured)
-//!
+//! * localhost (filter events originating from the local machine)
+//! * browser extensions (filter events caused by known problematic browser extensions)
+//! * web crawlers (filter events sent by user agents known to be web crawlers)
+//! * legacy browsers (filter events originating from legacy browsers, can be configured)
+#![warn(missing_docs)]
 
 use std::net::IpAddr;
 
-use crate::protocol::Event;
+use relay_general::protocol::Event;
 
 mod browser_extensions;
 mod client_ips;
@@ -18,21 +18,17 @@ mod common;
 mod config;
 mod csp;
 mod error_messages;
-
-#[cfg(feature = "uaparser")]
 mod legacy_browsers;
-#[cfg(feature = "uaparser")]
-mod web_crawlers;
-
 mod localhost;
 mod releases;
+mod web_crawlers;
 
 #[cfg(test)]
-mod test_utils;
+mod testutils;
 
-pub use common::*;
-pub use config::*;
-pub use csp::matches_any_origin;
+pub use crate::common::*;
+pub use crate::config::*;
+pub use crate::csp::matches_any_origin;
 
 /// Checks whether an event should be filtered for a particular configuration.
 ///
@@ -52,11 +48,8 @@ pub fn should_filter(
     error_messages::should_filter(event, &config.error_messages)?;
     localhost::should_filter(event, &config.localhost)?;
     browser_extensions::should_filter(event, &config.browser_extensions)?;
-    #[cfg(feature = "uaparser")]
-    {
-        legacy_browsers::should_filter(event, &config.legacy_browsers)?;
-        web_crawlers::should_filter(event, &config.web_crawlers)?;
-    }
+    legacy_browsers::should_filter(event, &config.legacy_browsers)?;
+    web_crawlers::should_filter(event, &config.web_crawlers)?;
 
     Ok(())
 }
