+++ filter/src/releases.rs
@@ -4,8 +4,9 @@
 //! (known old bad releases) and Sentry will ignore events originating from
 //! clients with the specified release.
 
-use crate::filter::{FilterStatKey, ReleasesFilterConfig};
-use crate::protocol::Event;
+use relay_general::protocol::Event;
+
+use crate::{FilterStatKey, ReleasesFilterConfig};
 
 /// Filters events generated by known problematic SDK clients.
 pub fn should_filter(event: &Event, config: &ReleasesFilterConfig) -> Result<(), FilterStatKey> {
@@ -21,9 +22,11 @@ pub fn should_filter(event: &Event, config: &ReleasesFilterConfig) -> Result<(),
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::filter::GlobPatterns;
-    use crate::protocol::{Event, LenientString};
-    use crate::types::Annotated;
+
+    use relay_general::protocol::{Event, LenientString};
+    use relay_general::types::Annotated;
+
+    use crate::GlobPatterns;
 
     fn get_event_for_release(release: &str) -> Event {
         Event {
