+++ general/Cargo.toml
@@ -14,9 +14,7 @@ cookie = { version = "0.12.0", features = ["percent-encode"] }
 debugid = { version = "0.7.0", features = ["serde"] }
 dynfmt = { version = "0.1.1", features = ["python", "curly"] }
 failure = "0.1.5"
-globset = "0.4.4"
 hmac = "0.7.1"
-ipnetwork = "0.14.0"
 itertools = "0.8.0"
 lazy_static = "1.3.0"
 lazycell = "1.2.1"
