+++ general/src/lib.rs
@@ -15,7 +15,6 @@ mod macros;
 #[macro_use]
 mod testutils;
 
-pub mod filter;
 pub mod pii;
 pub mod processor;
 pub mod protocol;
@@ -23,4 +22,4 @@ pub mod store;
 pub mod types;
 
 #[cfg(feature = "uaparser")]
-mod user_agent;
+pub mod user_agent;
