+++ server/src/actors/project.rs
@@ -21,7 +21,7 @@ use serde_json::Value;
 use url::Url;
 
 use relay_common::{metric, Config, LogError, ProjectId, PublicKey, RelayMode, RetryBackoff, Uuid};
-use relay_general::filter::{matches_any_origin, FiltersConfig};
+use relay_filter::{matches_any_origin, FiltersConfig};
 use relay_general::pii::{DataScrubbingConfig, PiiConfig};
 
 use crate::actors::outcome::DiscardReason;
